+++ Samples/sample04_dynamiccaltarget_PsychoPy.py
@@ -66,12 +66,11 @@ class FileWindow(wx.Frame):
 
 FileWindowValues = {}
 application = wx.App(False)
-fw = FileWindow(None,wx.ID_ANY,"Sample01_PsychoPy")
+fw = FileWindow(None,wx.ID_ANY,"Sample04_PsychoPy")
 application.MainLoop()
 
 
 dataFileName = FileWindowValues['filename']
-fp = open(dataFileName+'_local.csv','w')
 xy = FileWindowValues['imgsize'].split(',')
 cameraX = int(xy[0])
 cameraY = int(xy[1])
@@ -82,32 +81,29 @@ tracker.connect(FileWindowValues['address'])
 
 win = psychopy.visual.Window(size=(1024,768),units='pix')
 
-tracker.openDataFile(dataFileName+'.csv')
 tracker.sendSettings(GazeParser.config.getParametersAsDict())
 
-
 calarea = (-400,-300,400,300)
-"""calTargetPos = [[   0,   0],
+calTargetPos = [[   0,   0],
                 [-350,-250],[-350,  0],[-350,250],
                 [   0,-250],[   0,  0],[   0,250],
-                [ 350,-250],[ 350,  0],[ 350,250]]"""
-calTargetPos = ((   0,   0),
-                (-350,-250),(-350,  0),(-350,250),
-                (   0,-250),(   0,  0),(   0,250),
-                ( 350,-250),( 350,  0),( 350,250))
+                [ 350,-250],[ 350,  0],[ 350,250]]
 
 tracker.setCalibrationScreen(win)
 tracker.setCalibrationTargetPositions(calarea, calTargetPos)
-#tracker.setCalibrationTargetStimulus(psychopy.visual.Rect(win,width=100,height=100,units='pix'))
-calstim = [psychopy.visual.Rect(win,width=5,height=5,units='pix',lineColor=(1,0,0)),psychopy.visual.Rect(win,width=5,height=5,units='pix')]
+
+calstim = [psychopy.visual.Rect(win,width=4,height=4,units='pix',lineColor=(1,1,1),fillColor=(1,1,1)),
+           psychopy.visual.Rect(win,width=2,height=2,units='pix',lineColor=(-1,-1,-1),fillColor=(-1,-1,-1))]
 tracker.setCalibrationTargetStimulus(calstim)
 
 def callback(self, t, index, targetPos, currentPos):
-    if t<1.0:
-        self.caltarget[0].setSize(((10-9*t)*10,(10-9*t)*10))
+    if index==0:
+        return
     else:
-        self.caltarget[0].setSize((10,10))
-    print (t,targetPos,currentPos)
+        if t<1.0:
+            self.caltarget[0].setSize(((10-9*t)*4,(10-9*t)*4))
+        else:
+            self.caltarget[0].setSize((4,4))
 
 type(tracker).updateCalibrationTargetStimulusCallBack = callback
 
@@ -117,77 +113,4 @@ while True:
         sys.exit(0)
     if tracker.isCalibrationFinished():
         break
-
-stim = psychopy.visual.Rect(win, width=5, height=5, units='pix')
-marker = psychopy.visual.Rect(win, width=2, height=2, units='pix', fillColor=(1,1,0),lineWidth=0.1)
-
-trialClock = psychopy.core.Clock()
-for tr in range(2):
-    error = tracker.getSpatialError(message='Press space key')
-    
-    targetPositionList = [(100*random.randint(-3,3),100*random.randint(-3,3)) for i in range(10)]
-    targetPositionList.insert(0,(0,0))
-    currentPosition = 0
-    previousPosition = 0
-    stim.setPos(targetPositionList[currentPosition])
-    marker.setPos(targetPositionList[currentPosition])
-    
-    waitkeypress = True
-    while waitkeypress:
-        if 'space' in psychopy.event.getKeys():
-            waitkeypress = False
-        
-        stim.draw()
-        win.flip()
-
-    tracker.startRecording(message='trial'+str(tr+1))
-    tracker.sendMessage('STIM %s %s'%targetPositionList[currentPosition])
-    
-    data = []
-    trialClock.reset()
-    while True: 
-        currentTime = trialClock.getTime()
-        currentPosition = int(currentTime)
-        if currentPosition>=len(targetPositionList):
-            break
-        targetPosition = targetPositionList[currentPosition]
-        if previousPosition != currentPosition:
-            tracker.sendMessage('STIM %s %s'%targetPosition)
-            previousPosition = currentPosition
-        
-        preGet = trialClock.getTime()
-        eyePos= tracker.getEyePosition()
-        postGet = trialClock.getTime()
-        if not eyePos[0] == None:
-            data.append((1000*preGet,1000*postGet,1000*(postGet-preGet),
-                         targetPosition[0],targetPosition[1],eyePos[0],eyePos[1]))
-            marker.setPos((eyePos[0],eyePos[1]))
-        else:
-            data.append((1000*preGet,1000*postGet,1000*(postGet-preget),
-                         targetPosition[0],targetPosition[1],-65536,-65536))
-        
-        keyList = psychopy.event.getKeys()
-        if 'space' in keyList:
-            tracker.sendMessage('press space')
-        
-        stim.setPos(targetPosition)
-        stim.draw()
-        marker.draw()
-        win.flip()
-        
-    tracker.stopRecording(message='end trial')
-    
-    fp.write('trial%d\n' % (tr+1))
-    if error[0] != None:
-        fp.write('getSpatialError: %.2f,%.2f,%.2f\n' % (error[0],error[-1][0],error[-1][1]))
-    else:
-        fp.write('getSpatialError: None\n')
-    fp.write('SentAt,ReceivedAt,Lag,TargetX,TargetY,EyeX,EyeY\n')
-    for d in data:
-        fp.write('%.1f,%.1f,%.1f,%d,%d,%.1f,%.1f\n' % d)
-    fp.flush()
-    
-tracker.closeDataFile()
-
-fp.close()